Overhaul to how Room class works and how Dungeon class uses Rooms
This change was made to enable Dungeon class to hold many different types of rooms in a single container using base class pointers.

Room class now lets inherited classes do the construction of different types of rooms. Currently there is a SquareRoom and DiamondRoom class that makes different room templates. StartingRoom now inherits from SquareRoom.

DiamondRoom has a chance to contain an altar 'A'. Added a virtual hasAltar() function to Room and DiamondRoom to be prevent a room spawning with both chests and an altar.